+++ src/editor/core/command/CommandAdapt.ts
@@ -121,6 +121,7 @@ import {
   IGetAreaValueOption,
   IGetAreaValueResult,
   IInsertAreaOption,
+  ILocationAreaOption,
   ISetAreaPropertiesOption
 } from '../../interface/Area'
 import { IAreaBadge, IBadge } from '../../interface/Badge'
@@ -1287,17 +1288,12 @@ export class CommandAdapt {
     if (isDisabled) return null
     const { startIndex, endIndex } = this.range.getRange()
     if (!~startIndex && !~endIndex) return null
-    const { id, conceptId, value, width, height, imgDisplay } = payload
-    const imageId = id || getUUID()
+    const imageId = payload.id || getUUID()
     this.insertElementList([
       {
-        value,
-        width,
-        height,
-        conceptId,
+        ...payload,
         id: imageId,
-        type: ElementType.IMAGE,
-        imgDisplay
+        type: ElementType.IMAGE
       }
     ])
     return imageId
@@ -1990,6 +1986,10 @@ export class CommandAdapt {
     }
   }
 
+  public translate(path: string): string {
+    return this.i18n.t(path)
+  }
+
   public setLocale(payload: string) {
     this.i18n.setLocale(payload)
   }
@@ -2524,8 +2524,8 @@ export class CommandAdapt {
     this.draw.getArea().setAreaProperties(payload)
   }
 
-  public locationArea(areaId: string) {
-    const context = this.draw.getArea().getContextByAreaId(areaId)
+  public locationArea(areaId: string, options?: ILocationAreaOption) {
+    const context = this.draw.getArea().getContextByAreaId(areaId, options)
     if (!context) return
     const {
       range: { endIndex },