+++ src/editor/core/event/handlers/composition.ts
@@ -20,11 +20,9 @@ function compositionend(host: CanvasEvent, evt: CompositionEvent) {
     })
   } else {
     // 存在值：无法触发input事件需手动检测并触发渲染
-    setTimeout(() => {
-      if (host.compositionInfo) {
-        input(evt.data, host)
-      }
-    }, 1) // 如果为0，火狐浏览器会在input事件之前执行导致重复输入
+    if (host.compositionInfo) {
+      input(evt.data, host)
+    }
   }
   // 移除代理输入框数据
   const cursor = draw.getCursor()